last

+++ src/main/java/tn/esprit/spring/services/UserDetailsImpl.java
@@ -19,9 +19,9 @@ public class UserDetailsImpl implements UserDetails {
     @JsonIgnore
     private String password;
     private Collection<? extends GrantedAuthority> authorities;
-    public UserDetailsImpl(Long id, String username, String email, String password,
+    public UserDetailsImpl(Integer id, String username, String email, String password,
                            Collection<? extends GrantedAuthority> authorities) {
-        this.id = id;
+        this.id = Long.valueOf(id);
         this.username = username;
         this.email = email;
         this.password = password;
